Typed program values: int/bool as real Rust types (design) (#41)
* Add design spec for typed program int/bool/float values.

Documents tagged-union CpuStepResult, silent coercion, integer
division, and unchanged animation wire format.


* Add implementation plan for typed program values.

Breaks work into CpuStepResult enum, AST literals, eval,
sim wire mapping, and help/docs.


* Implement typed program values as bool/i64/f64.

Replace f64-tagged CpuStepResult with a real enum; coerce on
declare/assign; use integer division for int/int; keep animation
wire format as {k,v:number}. Update help for conversion rules.


* Remove completed typed-values implementation plan.

Keep the design spec; the plan was only scaffolding for the change.


---------

+++ robominer-sim/src/simulation/program_bridge.rs
@@ -250,7 +250,7 @@ impl Simulation {
                     // StartScan returns scan_time synchronously on issue (unlike move/mine).
                     let _ = step_result;
                     let scan_time = self.start_scan(robot_index, direction);
-                    let result = CpuStepResult::int_value(scan_time as f64);
+                    let result = CpuStepResult::int_value(i64::from(scan_time));
                     push_recorded_cpu_step(
                         &mut cpu_steps,
                         step_span,
